# perf: Remove dependency and improve performance