--- a/lib/calc-text-width.js
+++ b/lib/calc-text-width.js
@@ -1,5 +1,4 @@
 const widthsVerdana11 = require('./widths-verdana-11.json')
-const astralRegex = require('unicode-astral-regex')
 
 const SCALE = 10 // Prevent results like 60.599999999999994
 
@@ -8,7 +7,7 @@ const calcWidth = (charWidthTable) => {
   widthTable[64] = widthTable[64] + 6 // Slightly increase width of "@" by 0.6px
 
   return (text, astral) => {
-    if (astral) text = text.match(astralRegex) || []
+    if (astral) text = [...text]
 
     let total = 0
     let code = 0