Use the new spread operator to split the array therefore removing the need for the 'unicode-astral-regex' dependency.

All tests return a success and benchmark show a nice perfs improvement.

```
[classic] style, long params x 1,200,390 ops/sec ±4.84% (77 runs sampled)
[classic] style, full params x 1,774,494 ops/sec ±1.70% (86 runs sampled)
   [flat] style, long params x 651,913 ops/sec ±1.43% (88 runs sampled)
   [flat] style, full params x 636,148 ops/sec ±1.10% (86 runs sampled)
[classic] style, with emoji  x 665,623 ops/sec ±3.13% (85 runs sampled)
[classic] style, with icon   x 1,333,502 ops/sec ±0.79% (90 runs sampled)
   [flat] style, with emoji  x 390,781 ops/sec ±0.76% (90 runs sampled)
   [flat] style, with icon   x 544,615 ops/sec ±0.68% (89 runs sampled)

[classic] style, long params x 1,476,416 ops/sec ±0.56% (90 runs sampled)
[classic] style, full params x 1,825,783 ops/sec ±1.44% (89 runs sampled)
   [flat] style, long params x 684,052 ops/sec ±1.06% (88 runs sampled)
   [flat] style, full params x 644,841 ops/sec ±1.50% (90 runs sampled)
[classic] style, with emoji  x 1,057,921 ops/sec ±0.87% (88 runs sampled) => +159%
[classic] style, with icon   x 1,295,739 ops/sec ±3.06% (83 runs sampled)
   [flat] style, with emoji  x 514,076 ops/sec ±1.71% (87 runs sampled) => 131%
   [flat] style, with icon   x 564,612 ops/sec ±1.05% (88 runs sampled)
```

Only drawback is that it require support for the spread operator (...), supported since node 5.